Create deepCompare module
