+++ src/modules/deep-comparison/jest.config.js
@@ -1,0 +1,8 @@
+module.exports = {
+  roots: ["<rootDir>/src"],
+  testMatch: ["**/__tests__/**/*.+(ts|tsx|js)", "**/?(*.)+(spec|test).+(ts|tsx|js)"],
+  transform: {
+    "^.+\\.(ts|tsx)$": "ts-jest",
+  },
+  verbose: true,
+};
