+++ src/modules/deep-comparison/package.json
@@ -1,0 +1,35 @@
+{
+  "name": "@openmsupply/deep-comparison",
+  "version": "1.0.0",
+  "description": "Deep comparison of objects for equality, to arbitrary depth",
+  "main": "lib/deepCompare.js",
+  "types": "lib/deepCompare.d.ts",
+  "author": "Sustainable Solutions",
+  "license": "MIT",
+  "private": false,
+  "publishConfig": {
+    "registry": "https://npm.pkg.github.com/"
+  },
+  "repository": {
+    "type": "git",
+    "url": "ssh://git@github.com/openmsupply/application-manager-server.git",
+    "directory": "packages/deep-comparison"
+  },
+  "devDependencies": {
+    "@types/jest": "^26.0.10",
+    "@types/node": "^14.6.0",
+    "jest": "^26.4.0",
+    "rimraf": "^3.0.2",
+    "ts-jest": "^26.2.0",
+    "ts-node": "^8.10.2",
+    "typescript": "^3.9.7"
+  },
+  "scripts": {
+    "build": "rimraf ./lib && tsc",
+    "test": "jest"
+  },
+  "dependencies": {},
+  "files": [
+    "lib/**/*"
+  ]
+}
